# api/openapi-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Windshift API v2",
    "version": "2.0.0",
    "description": "Canonical API v2. Browser clients use the session-authenticated /api/v2 mount; non-browser clients use this bearer-authenticated mount."
  },
  "servers": [
    {"url": "/rest/api/v2"}
  ],
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get the authenticated user",
        "security": [{"BearerAuth": ["users:read"]}],
        "responses": {
          "200": {
            "description": "Authenticated user",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/UserDocument"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/AuthenticationError"},
          "403": {"$ref": "#/components/responses/PermissionError"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "crw_*"
      }
    },
...
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "email", "username", "first_name", "last_name", "full_name", "is_active", "is_agent", "agent_presence", "avatar_url", "timezone", "language", "created_at"],
        "properties": {
          "id": {"type": "integer"},
          "email": {"type": "string"},
          "username": {"type": "string"},
          "first_name": {"type": "string"},
          "last_name": {"type": "string"},
          "full_name": {"type": "string"},
          "is_active": {"type": "boolean"},
          "is_agent": {"type": "boolean"},
          "agent_presence": {"type": "string"},
          "avatar_url": {"type": "string", "nullable": true},
          "timezone": {"type": "string"},
          "language": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "UserDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {"data": {"$ref": "#/components/schemas/User"}}
      },
      "ErrorContent": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "details": {}
        }
      },
      "ErrorDocument": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": {"$ref": "#/components/schemas/ErrorContent"},
          "request_id": {"type": "string"}
        }
      },
...
    },
    "responses": {
...
      "AuthenticationError": {
        "description": "Authentication required or invalid",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorDocument"}}}
      },
      "PermissionError": {
        "description": "Required token scope is missing",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorDocument"}}}
      },
...
      "InternalError": {
        "description": "Unexpected server error",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorDocument"}}}
      }
    }
  }
}